+++ CSharpHTTPClient/Client.cs
@@ -14,7 +14,7 @@ using System.Threading;
 namespace SendGrid.CSharp.HTTP.Client
 {
     public class Response
-    {
+    {      
         public HttpStatusCode StatusCode;
         public HttpContent Body;
         public HttpResponseHeaders Headers;
@@ -64,17 +64,21 @@ namespace SendGrid.CSharp.HTTP.Client
 
     public class Client : DynamicObject
     {
+        private static HttpClient _httpClient = new HttpClient();
         public string Host;
         public Dictionary <string,string> RequestHeaders;
         public string Version;
         public string UrlPath;
         public string MediaType;
         public WebProxy WebProxy;
+        public TimeSpan Timeout;
+
         public enum Methods
         {
             DELETE, GET, PATCH, POST, PUT
         }
 
+        private int TimeoutDefault = 10;
 
         /// <summary>
         ///     REST API client.
@@ -90,6 +94,7 @@ namespace SendGrid.CSharp.HTTP.Client
             WebProxy = webProxy;
         }
 
+
         /// <summary>
         ///     REST API client.
         /// </summary>
@@ -97,8 +102,9 @@ namespace SendGrid.CSharp.HTTP.Client
         /// <param name="requestHeaders">A dictionary of request headers</param>
         /// <param name="version">API version, override AddVersion to customize</param>
         /// <param name="urlPath">Path to endpoint (e.g. /path/to/endpoint)</param>
+        /// <param name="timeOut">Set an Timeout parameter for the HTTP Client</param>
         /// <returns>Fluent interface to a REST API</returns>
-        public Client(string host, Dictionary<string,string> requestHeaders = null, string version = null, string urlPath = null)
+        public Client(string host, Dictionary<string,string> requestHeaders = null, string version = null, string urlPath = null, TimeSpan? timeOut = null)
         {
             Host = host;
             if(requestHeaders != null)
@@ -107,6 +113,7 @@ namespace SendGrid.CSharp.HTTP.Client
             }
             Version = (version != null) ? version : null;
             UrlPath = (urlPath != null) ? urlPath : null;
+            Timeout = (timeOut != null) ? (TimeSpan)timeOut : TimeSpan.FromSeconds(TimeoutDefault);
         }
 
         /// <summary>
@@ -172,11 +179,10 @@ namespace SendGrid.CSharp.HTTP.Client
             }
 
             UrlPath = null; // Reset the current object's state before we return a new one
-            return new Client(Host, RequestHeaders, Version, endpoint);
+            return new Client(Host, RequestHeaders, Version, endpoint, Timeout);
 
         }
 
-        /// <summary>
         /// Factory method to return the right HttpClient settings.
         /// </summary>
         /// <returns>Instance of HttpClient</returns>
@@ -195,8 +201,10 @@ namespace SendGrid.CSharp.HTTP.Client
                 return new HttpClient(httpClientHandler);
             }
 
-            return new HttpClient();
+            return _httpClient;
         }
+
+        /// <summary>
 
         /// <summary>
         ///     Add the authorization header, override to customize
@@ -306,7 +314,8 @@ namespace SendGrid.CSharp.HTTP.Client
         /// <returns>Response object</returns>
         public async virtual Task<Response> MakeRequest(HttpClient client, HttpRequestMessage request, CancellationToken cancellationToken)
         {
-            HttpResponseMessage response = await client.SendAsync(request, cancellationToken).ConfigureAwait(false);
+
+            HttpResponseMessage response = await client.SendAsync(request).ConfigureAwait(false);
             return new Response(response.StatusCode, response.Content, response.Headers);
         }
 
@@ -318,15 +327,17 @@ namespace SendGrid.CSharp.HTTP.Client
         /// <param name="requestBody">JSON formatted string</param>
         /// <param name="queryParams">JSON formatted queary paramaters</param>
         /// <returns>Response object</returns>
-        private async Task<Response> RequestAsync(string method, CancellationToken cancellationToken, string requestBody = null, string queryParams = null)
+        private async Task<Response> RequestAsync(string method, String requestBody = null, String queryParams = null, CancellationToken cancellationToken = null)
         {
-            using (var client = BuildHttpClient())
+            using (var client = new HttpClient())
             {
                 try
                 {
                     // Build the URL
                     client.BaseAddress = new Uri(Host);
+                    client.Timeout = Timeout;
                     string endpoint = BuildUrl(queryParams);
+
 
                     // Build the request headers
                     client.DefaultRequestHeaders.Accept.Clear();