+++ CHANGELOG.md
@@ -3,6 +3,16 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.2.0] - 2017-04-11
+### Added
+- #23: Timeout Parameter
+- Big thanks to [PandaBoy00](https://github.com/PandaBoy00) for the pull request!
+
+## [3.1.0] - 2017-03-01
+### Added
+- [PR #18](https://github.com/sendgrid/csharp-http-client/pull/18): Cache default httpclient
+- Big thanks to [Niels Timmermans](https://github.com/nillis) for the pull request!
+
 ## [3.0.0] - 2016-07-22
 ### BREAKING CHANGE
 - While your code may continue to work as before, the async behavior has changed, as we don't block on .Result anymore 
@@ -41,3 +51,4 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [1.0.2] - 2016-03-17
 ### Added
 - We are live!
+